pyproject.toml
[project]
name = "fastapi-sandbox"
version = "1.0.0"
description = "Amazon-Rakuten Product Research Tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.3",
    "fastapi>=0.128.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0",
    "uvicorn>=0.40.0",
    "pydantic-settings>=2.0",
    "redis>=5.0",
    "rq>=1.16",
    "jinja2>=3.1",
    "aiofiles>=24.0",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
    "python-amazon-sp-api>=0.18",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "ruff>=0.14.10",
]
